+++ NEWS.md
@@ -1,3 +1,10 @@
+# natmanager 0.4.8
+
+* Update to new style GitHub PAT
+* Recommend R > 4.0.2 for `natverse`
+* catch error in some situations when no GitHub PAT is set
+* moved CI to GitHub actions
+
 # natmanager 0.4.7
 
 * Bug fixes due to changes in `usethis` package resulting in deprecation of GitHub related APIs. Bump up version